package updater: replace --recheck with --redownload, making its behavior saner.
--redownload drops all 404 records thus triggering downloading of
the missing tarballs again on the next --traballs run.

This is better than the behavior of --recheck which simply ignored
404 for the current run since --tarballs run can be aborted and
resumed without losing any state, while --recheck had no idea which
tarballs were tried again and which were not. Also this allows triggering
redownloads asyncronously which is very convenient if you are running
nixpkgs-monitor as a service.